third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-request-basics.html

<!DOCTYPE html>
<title>Credential Manager: request() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined, "FIXME: We're currently always returning 'undefined'.");
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("request() should not reject, but did: " + reason.name);
}

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request().then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request({}).");
    t.step(function () {
        navigator.credentials.request({}).then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            federations: [ 'https://example.com/' ]
        }).then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            federations: [ 'https://example.com/' ],
            zeroClickOnly: true
        }).then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            zeroClickOnly: true
        }).then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());

(function() {
    var t = async_test("Verify the basics of request().");
    t.step(function () {
        navigator.credentials.request({
            notValid: 'yay!'
        }).then(
            t.step_func(stubResolverChecker.bind(t)),
            t.step_func(stubRejectionChecker.bind(t)));
    });
}());
</script>
